agent/rule/callback/monitor-http-status-code: add tests
