+++ internal/rule/callback/monitor-http-status-code_test.go
@@ -4,31 +4,88 @@
 
 package callback_test
 
-//func TestNewMonitorHTTPStatusCodeCallbacks(t *testing.T) {
-//	RunNativeCallbackTest(t, TestConfig{
-//		CallbacksCtor: callback.NewMonitorHTTPStatusCodeCallback,
-//		ExpectProlog:  true,
-//		PrologType:    reflect.TypeOf(callback.MonitorHTTPStatusCodePrologCallbackType(nil)),
-//		EpilogType:    reflect.TypeOf(callback.MonitorHTTPStatusCodeEpilogCallbackType(nil)),
-//		ValidTestCases: []ValidTestCase{
-//			{
-//				Rule: &NativeRuleContextMockup{},
-//				TestCallback: func(t *testing.T, rule *NativeRuleContextMockup, prolog sqhook.PrologCallback) {
-//					actualProlog, ok := prolog.(callback.MonitorHTTPStatusCodePrologCallbackType)
-//					require.True(t, ok)
-//					code := rand.Int()
-//					rule.On("AddMetricsValue", code, uint64(1)).Return().Once()
-//					epilog, err := actualProlog(nil, &code)
-//					// Check it behaves as expected
-//					require.NoError(t, err)
-//
-//					// Test the epilog if any
-//					if epilog != nil {
-//						require.True(t, ok)
-//						epilog()
-//					}
-//				},
-//			},
-//		},
-//	})
-//}
+import (
+	"net/http"
+	"testing"
+
+	"github.com/sqreen/go-agent/internal/event"
+	http_protection "github.com/sqreen/go-agent/internal/protection/http"
+	http_protection_mockups "github.com/sqreen/go-agent/internal/protection/http/_testlib/mockups"
+	"github.com/sqreen/go-agent/internal/protection/http/types"
+	"github.com/sqreen/go-agent/internal/rule/callback"
+	"github.com/sqreen/go-agent/internal/rule/callback/_testlib/mockups"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
+)
+
+func TestMonitorHTTPStatusCodeCallback(t *testing.T) {
+	t.Run("not 404", func(t *testing.T) {
+		r := &mockups.NativeRuleContextMockup{}
+		defer r.AssertExpectations(t)
+
+		r.ExpectPre(mock.Anything).Once()
+
+		cb, err := callback.NewMonitorHTTPStatusCodeCallback(r, nil)
+		require.NoError(t, err)
+
+		prolog := cb.(http_protection.ResponseMonitoringPrologCallbackType)
+		require.NotNil(t, prolog)
+
+		resp := &http_protection_mockups.ResponseMockup{}
+		defer resp.AssertExpectations(t)
+		expectedStatusCode := http.StatusExpectationFailed
+		resp.ExpectStatus().Return(expectedStatusCode).Once()
+
+		var respFace types.ResponseFace = resp
+		epilog, err := prolog(nil, &respFace)
+		require.NoError(t, err)
+		require.Nil(t, epilog)
+
+		r.AssertCalled(t, "Pre", mock.MatchedBy(func(cb func(callback.CallbackContext)) bool {
+			c := &mockups.CallbackContextMockup{}
+			defer c.AssertExpectations(t)
+			c.ExpectAddMetricsValue(expectedStatusCode, 1).Return(nil).Once()
+			cb(c)
+			return true
+		}))
+	})
+
+	t.Run("404", func(t *testing.T) {
+		r := &mockups.NativeRuleContextMockup{}
+		defer r.AssertExpectations(t)
+
+		r.ExpectPre(mock.Anything).Once()
+
+		cb, err := callback.NewMonitorHTTPStatusCodeCallback(r, nil)
+		require.NoError(t, err)
+
+		prolog := cb.(http_protection.ResponseMonitoringPrologCallbackType)
+		require.NotNil(t, prolog)
+
+		resp := &http_protection_mockups.ResponseMockup{}
+		defer resp.AssertExpectations(t)
+		expectedStatusCode := http.StatusNotFound
+		resp.ExpectStatus().Return(expectedStatusCode).Once()
+
+		var respFace types.ResponseFace = resp
+		epilog, err := prolog(nil, &respFace)
+		require.NoError(t, err)
+		require.Nil(t, epilog)
+
+		r.AssertCalled(t, "Pre", mock.MatchedBy(func(cb func(callback.CallbackContext)) bool {
+			c := &mockups.CallbackContextMockup{}
+			defer c.AssertExpectations(t)
+			c.ExpectAddMetricsValue(expectedStatusCode, 1).Return(nil).Once()
+			c.ExpectHandleAttack(false, mock.MatchedBy(func(opts []event.AttackEventOption) bool {
+				attack := &event.AttackEvent{}
+				for _, opt := range opts {
+					opt(attack)
+				}
+				return assert.Equal(t, &event.AttackEvent{Test: true, Info: struct{}{}}, attack)
+			})).Return(false).Once()
+			cb(c)
+			return true
+		}))
+	})
+}
